Update parseInt method to parseFloat

+++ script.js
@@ -51,19 +51,19 @@ function updateDisplay(btnValue, btnClass) {
   } else if (btnClass === "operator" && operator) {
     getOperator();
     display.value = operate(num1, num2, operator);
-    num1 = parseInt(display.value);
+    num1 = parseFloat(display.value);
     num2 = "";
     operator = btnValue;
   } else if (btnClass === "operator") {
     operator = btnValue;
-    num1 = parseInt(display.value);
+    num1 = parseFloat(display.value);
   } else if (num2) {
     display.value += btnValue;
-    num2 = parseInt(display.value);
+    num2 = parseFloat(display.value);
   } else if (operator) {
     display.value = "";
     display.value += btnValue;
-    num2 = parseInt(display.value);
+    num2 = parseFloat(display.value);
   } else if (!num1) {
     display.value = "";
     display.value += btnValue;
